+++ src/js/tools/time/daysInMonth.js
@@ -1,0 +1,3 @@
+export default function daysInMonth(month, year) {
+    return new Date(year, month, 0).getDate()
+}